requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0

# P1: ML/NLP Dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
umap-learn>=0.5.4
hdbscan>=0.8.33
keybert>=0.8.0
scikit-learn>=1.3.0
spacy>=3.7.0
numpy>=1.24.0
scipy>=1.11.0

